feat: push warmup method to streamer class
