+++ jina/serve/runtimes/gateway/websocket/gateway.py
@@ -108,7 +108,7 @@ class WebSocketGateway(BaseGateway):
             )
         )
 
-        self._warmup_task = asyncio.create_task(self.warmup())
+        self._warmup_task = asyncio.create_task(self.streamer.warmup())
         await self.server.setup()
 
     async def shutdown(self):
